requirements.txt
boto3
pdf2image
pytesseract
pillow